Made more changes to how invoice works (HTTP Changes)

+++ Models/Invoice.cs
@@ -10,7 +10,7 @@ namespace CheckIT.API.Models
         public DateTime InvoiceDate { get; set; }
         public bool OutgoingInv { get; set; }
         public bool IncomingInv { get; set; }
-        public float AmmountPaid { get; set; }
+        public decimal AmmountPaid { get; set; }
         public ICollection<Item> Items  { get; set; }
     }
 }
